third_party/WebKit/LayoutTests/fast/multicol/balance-float-with-margin-top-and-line-after-break-3.html

<!DOCTYPE html>
<p>There should be a cyan square in the bottom right corner of a blue rectangle.</p>
<div id="multicol" style="position:relative; columns:3; width:150px; line-height:20px; orphans:1; widows:1; background:blue;">
    <br>
    <br>
    <div id="elm" style="float:right; margin-top:10px; width:20px; background:cyan;">
        <br>
    </div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
    var multicol = document.getElementById("multicol");
    var elm = document.getElementById("elm");
    assert_equals(multicol.offsetHeight, 30);
    assert_equals(elm.offsetTop, 10);
}, "Float that should fit its top margin together with the border box.");
</script>
